email-app/backend/requirements.txt
# Core FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
pymongo==4.5.0
motor==3.3.2

# Data Validation
pydantic==2.4.2
pydantic-settings==2.0.3
email-validator==2.1.0
pytz>=2023.3           

# Environment & Config
python-dotenv==1.0.0

# HTTP Client
httpx==0.25.2

# Redis & Caching
redis>=4.5.2,<5.0.0
hiredis==2.2.3  # For better Redis performance

# Task Queue
celery[redis]==5.3.4
flower==2.0.1
kombu==5.3.4  # Celery dependency

# Authentication & Security
python-jose[cryptography]==3.3.0
bcrypt==4.1.2
passlib[bcrypt]==1.7.4
cryptography==41.0.7

# Templating
jinja2==3.1.2

# File Operations
aiofiles==23.2.1

# AWS (for SES)
boto3==1.34.13
botocore==1.34.13

# Async HTTP
aiohttp==3.9.1
aioredis==2.0.1  # For async Redis operations

# System Monitoring
psutil==5.9.6

# Date & Time
python-dateutil==2.8.2

# SMTP (if using custom SMTP)
aiosmtplib==3.0.1  # For async SMTP

# Data Processing (for CSV uploads)
pandas==2.1.4  # Optional but useful for bulk operations
openpyxl==3.1.2  # For Excel file support

# Logging & Monitoring
sentry-sdk==1.39.2  # Optional: For error tracking

